// src/core/modules/dao-api/models/UserDAOOwnershipResponse.ts
/**
 * DAO-API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: dev
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

/**
* Response schema for checking if a user owns a DAO
*/
export class UserDAOOwnershipResponse {
    /**
    * Whether the user owns a DAO
    */
    'hasDao'?: boolean;

    static readonly discriminator: string | undefined = undefined;

    static readonly mapping: {[index: string]: string} | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "hasDao",
            "baseName": "has_dao",
            "type": "boolean",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return UserDAOOwnershipResponse.attributeTypeMap;
    }

    public constructor() {
    }
}