// src/core/modules/dao-api/models/DAOModule.ts
/**
 * DAO-API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: dev
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

/**
* Module to add to or remove from a DAO
*/
export class DAOModule {
    'module': string;

    static readonly discriminator: string | undefined = undefined;

    static readonly mapping: {[index: string]: string} | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "module",
            "baseName": "module",
            "type": "string",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return DAOModule.attributeTypeMap;
    }

    public constructor() {
    }
}